spotify-oauth-gcp/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
google-cloud-secret-manager
jinja2
